Siirrä adapterit toiseen paikkaan
