+++ backend/src/projekti/adapter/adaptToAPI/index.ts
@@ -4,5 +4,4 @@ export * from "./adaptKasittelynTila";
 export * from "./adaptNahtavillaoloVaihe";
 export * from "./adaptSuunitteluSopimus";
 export * from "./adaptVuorovaikutusKierros";
-export * from "./adaptLogot";
 export * from "./adaptCommonToAPI";
